# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "How Fast Can You Eat?",
        "title": "CSH Marks",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/actions/group": {
            "post": {
                "description": "Gets a group of logged actions for auditing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Gets a group of logged actions",
                "parameters": [
                    {
                        "description": "Action Group",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetActionGroupInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetActionsGroupOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/archive": {
            "post": {
                "description": "Archives an Event to no longer be eligible for timing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Archives an Event",
                "parameters": [
                    {
                        "description": "Action Group",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArchiveEventInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/create": {
            "post": {
                "description": "Creates a new Event instance for timer sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Creates a New Event Instance",
                "parameters": [
                    {
                        "description": "Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNewEventInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/get": {
            "get": {
                "description": "Gets all open Events for standard users, or everything for Authorized Admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Gets All Authorized Events",
                "parameters": [
                    {
                        "description": "Action Group",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArchiveEventInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/open": {
            "post": {
                "description": "Opens an Event to be eligible for timing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Opens an Event",
                "parameters": [
                    {
                        "description": "Action Group",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OpenEventInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/leaderboard/delete": {
            "delete": {
                "description": "Deletes a User timer for a specific event, updating leaderboards accordingly",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Removes an Entry",
                "parameters": [
                    {
                        "description": "User / Event Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RemoveLeaderboardEntryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RemoveLeaderboardEntryOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/leaderboard/top": {
            "get": {
                "description": "Gets Data from a leaderboard database, supports pagination with the cursor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Gets Data for a Leaderboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetLeaderboardGroupOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/timer/reset": {
            "post": {
                "description": "Resets a User's Timer for the specified event. No-op if timer is not started",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timer"
                ],
                "summary": "Resets a User's Timer",
                "parameters": [
                    {
                        "description": "Event Id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetUserTimerInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/timer/start": {
            "post": {
                "description": "Starts a User's Timer for the specified event. No-op if already started",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timer"
                ],
                "summary": "Starts a User's Timer",
                "parameters": [
                    {
                        "description": "Event Id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StartUserTimerInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/timer/stop": {
            "post": {
                "description": "Stops a User's Timer for the specified event. No-op if timer is already stopped",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timer"
                ],
                "summary": "Stops a User's Timer",
                "parameters": [
                    {
                        "description": "Event Id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StopUserTimerInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.UnauthorizedResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ArchiveEventInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.CreateNewEventInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.GetActionGroupCursor": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.GetActionGroupInput": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.GetActionGroupPart": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "row_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.GetActionsGroupOutput": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetActionGroupPart"
                    }
                },
                "cursor": {
                    "$ref": "#/definitions/models.GetActionGroupCursor"
                }
            }
        },
        "models.GetLeaderboardEntry": {
            "type": "object",
            "properties": {
                "event_name": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.GetLeaderboardGroupOutput": {
            "type": "object",
            "properties": {
                "can_modify": {
                    "type": "boolean"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetLeaderboardEntry"
                    }
                }
            }
        },
        "models.GetLeaderboardTopOutput": {
            "type": "object",
            "properties": {
                "can_modify": {
                    "type": "boolean"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetLeaderboardEntry"
                    }
                }
            }
        },
        "models.OpenEventInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.RemoveLeaderboardEntryInput": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.RemoveLeaderboardEntryOutput": {
            "type": "object"
        },
        "models.ResetUserTimerInput": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                }
            }
        },
        "models.StartUserTimerInput": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                }
            }
        },
        "models.StopUserTimerInput": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                }
            }
        },
        "models.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}